Implement robust CVar-based configuration system
- Introduced `IConfigurationManager` and `ConfigurationManager` for type-safe settings.
- Added support for JSON persistence and change notification events.
- Created `GlobalSettings` to centralize graphics and server default values.
- Implemented in-game `SettingsPanel` (Client) and `ProjectSettingsPanel` (Editor) using reflection for dynamic UI generation.
- Integrated the new system into the Server, Client rendering pipeline, and Editor.
- Fixed 64-bit coordinate conversion issues in Editor components.
- Added unit tests for the configuration system.

+++ Developers/Editor/History/DeleteObjectCommand.cs
@@ -6,7 +6,7 @@ namespace Editor.History
     {
         private readonly IGameState _gameState;
         private readonly IGameObject _target;
-        private readonly int _x, _y, _z;
+        private readonly long _x, _y, _z;
 
         public string Name => $"Delete {_target.ObjectType.Name}";
