+++ Developers/Editor/History/MoveObjectCommand.cs
@@ -5,12 +5,12 @@ namespace Editor.History
     public class MoveObjectCommand : IUndoableCommand
     {
         private readonly IGameObject _target;
-        private readonly int _oldX, _oldY, _oldZ;
-        private readonly int _newX, _newY, _newZ;
+        private readonly long _oldX, _oldY, _oldZ;
+        private readonly long _newX, _newY, _newZ;
 
         public string Name => $"Move {_target.ObjectType.Name}";
 
-        public MoveObjectCommand(IGameObject target, int oldX, int oldY, int oldZ, int newX, int newY, int newZ)
+        public MoveObjectCommand(IGameObject target, long oldX, long oldY, long oldZ, long newX, long newY, long newZ)
         {
             _target = target;
             _oldX = oldX;
